removed printf warnings
[git-p4: depot-paths = "//tbl/": change = 339]

+++ Code/Modules/player/DemoPlayer.cpp
@@ -117,7 +117,7 @@ int BaseDemoPlayer::updateTextures()
 		}
 		else
 		{
-			printf("sceIoPollAsync returned %x with res = %x\n", ret, res);
+			printf("sceIoPollAsync returned %x with res = %x\n", ret, (unsigned)res);
 			m_currentLoad = 0;
 			return m_texQueue.size();
 		}
@@ -147,7 +147,7 @@ int BaseDemoPlayer::updateTextures()
 			sceIoLseek(m_currentLoad, 0, PSP_SEEK_SET);
 //			printf("file size = %i\n", size);
 			m_currentTexture = (data::MtxHeader*)malloc(size + sizeof(data::MtxHeader));
-			printf("mem ptr = %x\n", m_currentTexture);
+			printf("mem ptr = %x\n", (unsigned)m_currentTexture);
 			if (sceIoChangeAsyncPriority(m_currentLoad, 0x8))
 			{
 				printf("sceIoChangeAsyncPriority failed\n");
